Counts, capacities and shift amounts are unsigned
27 constants were i32 only because C's int was. Each was used almost
exclusively as `as usize` / `as u32` (or as a `0..N` bound whose loop
variable then only indexed), so the type was being contradicted at every
use. Found by flipping each signed constant to u32 and usize on a scratch
copy and counting the compile errors: these compile with no other change.

usize (counts, capacities, lengths):
  MAXBUTTONS MAXCEILINGS MAXPLATS MAXSWITCHES NUMCARDS NUMWEAPONS
  NUMPOWERS NUMMAPS NUMSFX NUM_VIRTUAL_BUTTONS MAXVISPLANES
  MAXVISSPRITES MAXDRAWSEGS MAXLIGHTZ SAVEGAMESIZE VERSIONSIZE TEXTWAIT
u32 (shift amounts, a value only ever used as u32):
  NET_MAXPLAYERS ANGLETOFINESHIFT ANGLETOSKYSHIFT HEIGHTBITS
  LIGHTSEGSHIFT LIGHTZSHIFT TRANSLATION_SHIFT BT_WEAPONSHIFT
  BTS_SAVESHIFT DEFAULT_SPECHIT_MAGIC

About 45 redundant casts go (clippy --fix). Two rewrites beyond that:
- draw_planes compared `ds_p as i64 > i64::from(MAXDRAWSEGS)` (and the
  same for visplanes); with usize that is `ds_p > MAXDRAWSEGS`.
- init_switch_list iterated 0..MAXSWITCHES only to index the table; it
  now walks the table with take(MAXSWITCHES).
- r_plane's light index is a usize clamp instead of u32 + casts.

Deliberately not touched: MAXPLAYERS (needs the player indices, still
i32 in ~30 places, retyped with it), SCREENWIDTH/HEIGHT and the
fixed-point/angle/offset constants (signed arithmetic, can go negative).

Verified: cargo test --release (57 pass, goldens unchanged), clippy and
rustfmt clean, whole workspace builds.

+++ rust-doomgeneric/engine/src/g_game.rs
@@ -375,7 +375,7 @@ pub const DEH_DEFAULT_INITIAL_BULLETS: i32 = 50;
 pub const DEH_INITIAL_HEALTH: i32 = DEH_DEFAULT_INITIAL_HEALTH;
 pub const DEH_INITIAL_BULLETS: i32 = DEH_DEFAULT_INITIAL_BULLETS;
 pub const DOOM_191_VERSION: i32 = 111;
-pub const SAVEGAMESIZE: i32 = 0x2c000;
+pub const SAVEGAMESIZE: usize = 0x2c000;
 pub const TURBOTHRESHOLD: i32 = 0x32;
 pub static ANGLETURN: [Fixed; 3] = [640, 1280, 320];
 static WEAPON_ORDER_TABLE: [WeaponOrder; 9] = [
@@ -1400,7 +1400,7 @@ pub fn do_save_game(state: &mut GameState) {
     archive_specials(&mut state.world);
     write_save_game_eof(&mut state.world.p_saveg);
     if state.game.g_game.vanilla_savegame_limit != 0
-        && state.world.p_saveg.save_buffer.len() > SAVEGAMESIZE as usize
+        && state.world.p_saveg.save_buffer.len() > SAVEGAMESIZE
     {
         error("Savegame buffer overrun");
     }
